# 1777 optimize decentralized script for speed