--- a/cea/optimization/preprocessing/decentralized_building_main.py
+++ b/cea/optimization/preprocessing/decentralized_building_main.py
@@ -30,12 +30,12 @@ def disconnected_building_main(locator, total_demand, config, prices, lca):
     buildings_name_with_space_heating = get_building_names_with_load(total_demand, load_name='Qhs_sys_MWhyr')
     buildings_name_with_cooling = get_building_names_with_load(total_demand, load_name='QC_sys_MWhyr')
 
-    if (buildings_name_with_heating != [] and buildings_name_with_space_heating != []):
+    if buildings_name_with_heating and buildings_name_with_space_heating:
         decentralized_buildings_heating.disconnected_buildings_heating_main(locator, total_demand,
                                                                             buildings_name_with_heating,
                                                                             config, prices, lca)
 
-    if buildings_name_with_cooling != []:
+    if buildings_name_with_cooling:
         decentralized_buildings_cooling.disconnected_buildings_cooling_main(locator,
                                                                             buildings_name_with_cooling,
                                                                             total_demand,