--- a/cea/optimization/preprocessing/decentralized_buildings_cooling.py
+++ b/cea/optimization/preprocessing/decentralized_buildings_cooling.py
@@ -542,14 +542,14 @@ def initialize_result_tables_for_supply_configurations(Qc_nom_SCU_W):
     return operation_results
 
 
-def calc_ACH_operation(T_ground_K, T_SC_hw_in_C, T_chw_re_K, T_chw_sup_K, chiller_prop, mdot_chw_kgpers, ACH_type):
-    chiller_prop = chiller_absorption.AbsorptionChiller(chiller_prop, ACH_type)
+def calc_ACH_operation(T_ground_K, T_SC_hw_in_C, T_chw_re_K, T_chw_sup_K, absorption_chiller, mdot_chw_kgpers, ACH_type):
+    absorption_chiller = chiller_absorption.AbsorptionChiller(absorption_chiller, ACH_type)
     SC_to_single_ACH_operation = np.vectorize(chiller_absorption.calc_chiller_main)(mdot_chw_kgpers,
                                                                                     T_chw_sup_K,
                                                                                     T_chw_re_K,
                                                                                     T_SC_hw_in_C,
                                                                                     T_ground_K,
-                                                                                    chiller_prop)
+                                                                                    absorption_chiller)
 
 
     el_ACH_Wh = np.asarray([x['wdot_W'] for x in SC_to_single_ACH_operation])