--- a/cea/optimization/slave/cooling_main.py
+++ b/cea/optimization/slave/cooling_main.py
@@ -9,6 +9,7 @@ from __future__ import division
 
 import numpy as np
 import pandas as pd
+import cea.inputlocator
 
 from cea.constants import HOURS_IN_YEAR
 from cea.optimization.constants import T_TANK_FULLY_DISCHARGED_K, DT_COOL, VCC_T_COOL_IN, ACH_T_IN_FROM_CHP_K
@@ -37,7 +38,7 @@ def district_cooling_network(locator,
     """
     Computes the parameters for the cooling of the complete DCN
 
-    :param locator: path to res folder
+    :param cea.inputlocator.InputLocator locator: path to res folder
     :param network_features: network features
     :param prices: Prices imported from the database
     :type locator: string
@@ -76,7 +77,7 @@ def district_cooling_network(locator,
         T_source_average_Lake_K = np.zeros(HOURS_IN_YEAR)
 
     # get properties of technology used in this script
-    ACH_prop = AbsorptionChiller(pd.read_excel(locator.get_supply_systems(), sheet_name="Absorption_chiller"), 'double')
+    absorption_chiller = AbsorptionChiller(pd.read_excel(locator.get_database_supply_systems(), sheet_name="Absorption_chiller"), 'double')
     CCGT_prop = calc_cop_CCGT(master_to_slave_variables.NG_Trigen_CCGT_size_W, ACH_T_IN_FROM_CHP_K, "NG")
 
     # intitalize variables
@@ -115,7 +116,7 @@ def district_cooling_network(locator,
                                                     daily_storage,
                                                     T_ground_K[hour],
                                                     master_to_slave_variables,
-                                                    ACH_prop,
+                                                    absorption_chiller,
                                                     CCGT_prop)
 
             Q_DailyStorage_gen_directload_W[hour] = thermal_output['Q_DailyStorage_gen_directload_W']