This PR addresses #1777 but actually goes a bit farther. The calculation in `cea.technologies.chiller_absorption.calc_operating_conditions` was replaced - the `sympy.linsolve` method call is not needed if you already have a symbolic solution to the equation. See the docstring for that method on how this was done. The end product is a really long and hairy expression but _so much faster_.

Execution speed went from:
- original: 2931.58s
- using expression: 226.28s
- smarter use of AbsorptionChiller (less pandas __getitem__ calls): ~180s
- added multiprocessing, 7 cores:  96.41s

I only worked on the cooling case. Do we need to do the same for heating?